requirements.txt
streamlit
requests
pandas
mysql-connector-python
python-dotenv
